Fixed file opening warningss. Replace QStringRef with QStringView and removed Qt5Compat dependency. Shape preview on ObjTools list browsing. CMakeLists system independent OpenAL linking and OpenGL linking.

+++ src/tsre/camera/Camera.h
@@ -57,7 +57,7 @@ public:
     virtual void setCameraObject(GameObj* o);
     virtual PreciseTileCoordinate* getCurrentPos();
     float * pozT = NULL;
-    int starex, starey;
+    float starex, starey;
     float fov;
 protected:
     float pos[3];
